requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
apache-airflow-providers-postgres
sqlalchemy